+++ regattaserver/kv.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/jamf/regatta/regattapb"
 	serrors "github.com/jamf/regatta/storage/errors"
+	"github.com/jamf/regatta/util/iter"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
@@ -52,6 +53,56 @@ func (s *KVServer) Range(ctx context.Context, req *regattapb.RangeRequest) (*reg
 		return nil, status.Error(codes.Internal, err.Error())
 	}
 	return val, nil
+}
+
+// IterateRange gets the keys in the range from the key-value store.
+func (s *KVServer) IterateRange(req *regattapb.RangeRequest, srv regattapb.KV_IterateRangeServer) error {
+	if req.GetLimit() < 0 {
+		return status.Errorf(codes.InvalidArgument, "limit must be a positive number")
+	} else if req.GetKeysOnly() && req.GetCountOnly() {
+		return status.Error(codes.InvalidArgument, "keys_only and count_only must not be set at the same time")
+	} else if req.GetMinModRevision() > 0 {
+		return status.Error(codes.Unimplemented, "min_mod_revision not implemented")
+	} else if req.GetMaxModRevision() > 0 {
+		return status.Error(codes.Unimplemented, "max_mod_revision not implemented")
+	} else if req.GetMinCreateRevision() > 0 {
+		return status.Error(codes.Unimplemented, "min_create_revision not implemented")
+	} else if req.GetMaxCreateRevision() > 0 {
+		return status.Error(codes.Unimplemented, "max_create_revision not implemented")
+	}
+
+	if len(req.GetTable()) == 0 {
+		return status.Error(codes.InvalidArgument, "table must be set")
+	}
+
+	if len(req.GetKey()) == 0 {
+		return status.Error(codes.InvalidArgument, "key must be set")
+	}
+
+	ctx := srv.Context()
+	r, err := s.Storage.IterateRange(ctx, req)
+	if err != nil {
+		if errors.Is(err, serrors.ErrTableNotFound) {
+			return status.Error(codes.NotFound, "table not found")
+		}
+		return status.Error(codes.Internal, err.Error())
+	}
+	pull, stop := iter.Pull(r)
+	defer stop()
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+		default:
+			response, valid := pull()
+			if !valid {
+				return nil
+			}
+			if err := srv.Send(response); err != nil {
+				return status.Error(codes.Internal, err.Error())
+			}
+		}
+	}
 }
 
 // Put implements proto/regatta.proto KV.Put method.
@@ -134,23 +185,8 @@ func (r *ReadonlyKVServer) DeleteRange(_ context.Context, _ *regattapb.DeleteRan
 // It is allowed to modify the same key several times within one txn (the result will be the last Op that modified the key).
 // Readonly transactions allowed using follower API.
 func (r *ReadonlyKVServer) Txn(ctx context.Context, req *regattapb.TxnRequest) (*regattapb.TxnResponse, error) {
-	if isReadonlyTransaction(req) {
+	if req.IsReadonly() {
 		return r.KVServer.Txn(ctx, req)
 	}
 	return nil, status.Error(codes.Unimplemented, "writable Txn not implemented for follower")
-}
-
-func isReadonlyTransaction(req *regattapb.TxnRequest) bool {
-	for _, op := range req.Success {
-		if _, ok := op.Request.(*regattapb.RequestOp_RequestRange); !ok {
-			return false
-		}
-	}
-
-	for _, op := range req.Failure {
-		if _, ok := op.Request.(*regattapb.RequestOp_RequestRange); !ok {
-			return false
-		}
-	}
-	return true
 }