+++ regattaserver/regattaserver_test.go
@@ -8,21 +8,29 @@ import (
 
 	"github.com/jamf/regatta/regattapb"
 	"github.com/jamf/regatta/storage/table"
+	"github.com/jamf/regatta/util/iter"
 )
 
 // MockStorage implements trivial storage for testing purposes.
 type MockStorage struct {
-	rangeResponse       regattapb.RangeResponse
-	putResponse         regattapb.PutResponse
-	deleteRangeResponse regattapb.DeleteRangeResponse
-	txnResponse         regattapb.TxnResponse
-	rangeError          error
-	putError            error
-	deleteError         error
+	rangeResponse        regattapb.RangeResponse
+	iterateRangeResponse iter.Seq[*regattapb.RangeResponse]
+	putResponse          regattapb.PutResponse
+	deleteRangeResponse  regattapb.DeleteRangeResponse
+	txnResponse          regattapb.TxnResponse
+	rangeError           error
+	iterateRangeError    error
+	putError             error
+	deleteError          error
+	txnError             error
 }
 
 func (s *MockStorage) Range(_ context.Context, _ *regattapb.RangeRequest) (*regattapb.RangeResponse, error) {
 	return &s.rangeResponse, s.rangeError
+}
+
+func (s *MockStorage) IterateRange(_ context.Context, _ *regattapb.RangeRequest) (iter.Seq[*regattapb.RangeResponse], error) {
+	return s.iterateRangeResponse, s.iterateRangeError
 }
 
 func (s *MockStorage) Put(_ context.Context, _ *regattapb.PutRequest) (*regattapb.PutResponse, error) {
@@ -34,7 +42,7 @@ func (s *MockStorage) Delete(_ context.Context, _ *regattapb.DeleteRangeRequest)
 }
 
 func (s *MockStorage) Txn(_ context.Context, _ *regattapb.TxnRequest) (*regattapb.TxnResponse, error) {
-	return &s.txnResponse, s.deleteError
+	return &s.txnResponse, s.txnError
 }
 
 type MockTableService struct {